requirements.txt
lxml_html_clean
fastapi
gnews
newspaper3k
openai
protobuf
python-dotenv
uvicorn
uvloop